Move `get_prev_next_pagination_links` to utils
This will mean it can later be reused whereever we want
